+++ my-app/src/index.js
@@ -7,9 +7,9 @@ import registerServiceWorker from './registerServiceWorker';
 
 
 ReactDOM.render((
-	<BrowserRouter >
-		<App />
-	</BrowserRouter>
-	), document.getElementById('root'))
+  <BrowserRouter >
+    <App />
+  </BrowserRouter>
+  ), document.getElementById('root'))
 
 registerServiceWorker();
